add swal default val

+++ client/app/routes/MyAccount/containers/ItemContainer.jsx
@@ -1,5 +1,7 @@
 import { connect } from 'react-redux';
-import { popupTwoButtons } from 'modules/popup';
+// import { popupTwoButtons } from 'modules/popup';
+import swal, { confirmConfig } from 'lib/swal';
+// TODO: move to layout
 
 import Container from '../components/Item';
 import { fetchItems, deleteItem, reset } from '../modules/myItem';
@@ -9,25 +11,23 @@ const mapStateToProps = (state) => {
   return ({ environment, mine, myItem });
 };
 
-const mapDispatchToProps = dispatch => ({
-  dispatch,
-  dispatchFetchItem: catId => dispatch(fetchItems(catId)),
-  dispatchReset: () => dispatch(reset()),
-  dispatchDelete: (pid) => {
-    const onAction = () => {
+const mapDispatchToProps = (dispatch) => {
+  const dispatchDelete = (pid) => {
+    // const { showAlert } = swal;
+    swal(confirmConfig({
+      title: '確定下架此商品？',
+      text: '下架之後即無法還原',
+      confirmButtonText: '下架',
+    })).then(() => {
       dispatch(deleteItem(pid));
-    };
-    dispatch(popupTwoButtons(
-      {
-        title: '確定下架此商品？',
-        content: '下架之後即無法還原',
-        rightBtnContent: '下架',
-        onRightClick: onAction,
-        leftBtnContent: '取消',
-        onLeftClick: () => {},
-      },
-    ));
-  }
-});
+    }, () => {});
+  };
+  return ({
+    dispatch,
+    dispatchFetchItem: catId => dispatch(fetchItems(catId)),
+    dispatchReset: () => dispatch(reset()),
+    dispatchDelete,
+  });
+};
 
 export default connect(mapStateToProps, mapDispatchToProps)(Container);
